Wayland display server
Decide which display server to run based on the configuration, but
always fallback to X11 if settings are wrong.

Run the greeter according to the display server in use and use xcb
for keyboard only when appropriate.

While we are here we make sure that the greeter is using the
correct QPA plugin on X11, as a matter of fact we used to rely
on the default which usually happens to be xcb.

Since the greeter on Wayland can be started only when the compositor
is up and running, the helper startup is now asynchronous.
After the PAM authentication and process execution the helper waits
for a the sessionStarted(bool) signal from UserSession; it is
delivered right after the process is executed for X11 sessions,
while it is delivered as soon as the Wayland socket is created for
Wayland display server sessions but a timeout make sure the helper
doesn't get stuck waiting for the socket to be created.